+++ html/requirements/RequirementsChecker.php
@@ -12,7 +12,7 @@ if (version_compare(PHP_VERSION, '4.3', '<')) {
 
 /**
  * The Craft Requirement Checker allows checking if the current system meets the minimum requirements for running a
- * Craft 3 application.
+ * Craft 4 application.
  *
  * This class allows rendering of the requirement report through a web browser or command line interface.
  *
@@ -50,8 +50,9 @@ class RequirementsChecker
 
     var $result;
 
-    var $requiredMySqlVersion = '5.5.0';
-    var $requiredPgSqlVersion = '9.5';
+    var $requiredMySqlVersion = '5.7.8';
+    var $requiredMariaDbVersion = '10.2.7';
+    var $requiredPgSqlVersion = '10.0';
 
     /**
      * Check the given requirements, collecting results into internal field.
@@ -385,7 +386,7 @@ class RequirementsChecker
 
         if ($conn === null) {
             try {
-                $conn = new PDO($this->dsn, $this->dbUser, $this->dbPassword);
+                $conn = new PDO($this->dsn ?? '', $this->dbUser, $this->dbPassword);
                 $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
             } catch (PDOException $e) {
                 $conn = false;
@@ -427,6 +428,27 @@ class RequirementsChecker
     }
 
     /**
+     * This method attempts to see if MySQL timezone data has been populated on
+     * the MySQL server Craft is configured to use.
+     *
+     * https://dev.mysql.com/doc/refman/5.7/en/time-zone-support.html
+     *
+     * @param PDO $conn
+     * @return bool
+     */
+    function validateDatabaseTimezoneSupport($conn)
+    {
+        $query = $conn->query("SELECT CONVERT_TZ('2007-03-11 2:00:00','US/Eastern','US/Central') AS time1");
+        $result = $query->fetchColumn();
+
+        if (!$result) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
      * @return array
      */
     function iniSetRequirement()
@@ -451,20 +473,20 @@ class RequirementsChecker
         // ini_set can return false or an empty string depending on your php version / FastCGI.
         // If ini_set has been disabled in php.ini, the value will be null because of our muted error handler
         if ($result === null) {
-            $memo = 'It looks like <a rel="noopener" target="_blank" href="http://php.net/manual/en/function.ini-set.php">ini_set</a> has been disabled in your <code>php.ini</code> file. Craft requires that to operate.';
+            $memo = 'It looks like <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.ini-set.php">ini_set</a> has been disabled in your <code>php.ini</code> file. Craft requires that to operate.';
             $condition = false;
         }
 
         // ini_set can return false or an empty string or the current value of memory_limit depending on your php
         // version and FastCGI. Regard, calling it didn't work, but there was no error.
         else if ($result === false || $result === '' || $result === $newValue) {
-            $memo = 'It appears calls to <a rel="noopener" target="_blank" href="http://php.net/manual/en/function.ini-set.php">ini_set</a> are not working for Craft. You may need to increase some settings in your php.ini file such as <a rel="noopener" target="_blank" href="http://php.net/manual/en/ini.core.php#ini.memory-limit">memory_limit</a> and <a rel="noopener" target="_blank" href="http://php.net/manual/en/info.configuration.php#ini.max-execution-time">max_execution_time</a> for long running operations like updating and asset transformations.';
+            $memo = 'It appears calls to <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.ini-set.php">ini_set</a> are not working for Craft. You may need to increase some settings in your php.ini file such as <a rel="noopener" target="_blank" href="https://php.net/manual/en/ini.core.php#ini.memory-limit">memory_limit</a> and <a rel="noopener" target="_blank" href="https://php.net/manual/en/info.configuration.php#ini.max-execution-time">max_execution_time</a> for long running operations like updating and asset transformations.';
 
             // Set mandatory to false here so it's not a "fatal" error, but will be treated as a warning.
             $mandatory = false;
             $condition = false;
         } else {
-            $memo = 'Calls to <a rel="noopener" target="_blank" href="http://php.net/manual/en/function.ini-set.php">ini_set</a> are working correctly.';
+            $memo = 'Calls to <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.ini-set.php">ini_set</a> are working correctly.';
             $condition = true;
         }
 
@@ -479,7 +501,7 @@ class RequirementsChecker
     /**
      * @return array
      *
-     * @see http://php.net/manual/en/ini.core.php#ini.memory-limit
+     * @see https://php.net/manual/en/ini.core.php#ini.memory-limit
      */
     function memoryLimitRequirement()
     {
@@ -500,7 +522,7 @@ class RequirementsChecker
     /**
      * @return array
      *
-     * @see http://php.net/manual/en/info.configuration.php#ini.max-execution-time
+     * @see https://php.net/manual/en/info.configuration.php#ini.max-execution-time
      */
     function maxExecutionTimeRequirement()
     {
@@ -513,6 +535,28 @@ class RequirementsChecker
             'name' => 'Max Execution Time',
             'mandatory' => false,
             'condition' => $maxExecutionTime === 0 || $maxExecutionTime >= 120,
+            'memo' => $memo,
+        );
+    }
+
+    /**
+     * @return array
+     */
+    function webAliasRequirement()
+    {
+        $aliases = Craft::$app->getConfig()->getGeneral()->aliases;
+        $memo = 'We recommend explicitly overriding the <a rel="noopener" target="_blank" href="https://craftcms.com/docs/3.x/config/#aliases">@web alias</a>.';
+        $pass = false;
+
+        if (isset($aliases['web']) || isset($aliases['@web'])) {
+            $memo = 'Your @web alias is set correctly';
+            $pass = true;
+        }
+
+        return array(
+            'name' => 'Ensure @web alias is explicitly overridden',
+            'mandatory' => false,
+            'condition' => $pass,
             'memo' => $memo,
         );
     }