+++ html/requirements/requirements.php
@@ -6,10 +6,10 @@
 /** @var RequirementsChecker $this */
 $requirements = array(
     array(
-        'name' => 'PHP 7.4+',
+        'name' => 'PHP 8.0.2+',
         'mandatory' => true,
-        'condition' => PHP_VERSION_ID >= 70400,
-        'memo' => 'PHP 7.4 or later is required.',
+        'condition' => PHP_VERSION_ID >= 80002,
+        'memo' => 'PHP 8.0.2 or later is required.',
     ),
 );
 
@@ -21,20 +21,37 @@ switch ($this->dbDriver) {
             'name' => 'PDO MySQL extension',
             'mandatory' => true,
             'condition' => extension_loaded('pdo_mysql'),
-            'memo' => 'The <http://php.net/manual/en/ref.pdo-mysql.php>PDO MySQL</a> extension is required.'
+            'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/ref.pdo-mysql.php">PDO MySQL</a> extension is required.'
         );
         if ($conn !== false) {
+            $version = $conn->getAttribute(PDO::ATTR_SERVER_VERSION);
+            if (preg_match('/[\d.]+-([\d.]+)-\bMariaDB\b/', $version, $match)) {
+                $name = 'MariaDB';
+                $version = $match[1];
+                $requiredVersion = $this->requiredMariaDbVersion;
+                $tzUrl = 'https://mariadb.com/kb/en/time-zones/#mysql-time-zone-tables';
+            } else {
+                $name = 'MySQL';
+                $requiredVersion = $this->requiredMySqlVersion;
+                $tzUrl = 'https://dev.mysql.com/doc/refman/5.7/en/time-zone-support.html';
+            }
             $requirements[] = array(
-                'name' => "MySQL {$this->requiredMySqlVersion}+",
+                'name' => "{$name} {$requiredVersion}+",
                 'mandatory' => true,
-                'condition' => $this->checkDatabaseServerVersion($conn, $this->requiredMySqlVersion),
-                'memo' => "MySQL {$this->requiredMySqlVersion} or higher is required to run Craft CMS.",
+                'condition' => version_compare($version, $requiredVersion, '>='),
+                'memo' => "{$name} {$requiredVersion} or higher is required to run Craft CMS.",
             );
             $requirements[] = array(
-                'name' => 'MySQL InnoDB support',
+                'name' => "{$name} InnoDB support",
                 'mandatory' => true,
                 'condition' => $this->isInnoDbSupported($conn),
-                'memo' => 'Craft CMS requires the MySQL InnoDB storage engine to run.',
+                'memo' => "Craft CMS requires the {$name} InnoDB storage engine to run.",
+            );
+            $requirements[] = array(
+                'name' => "{$name} timezone support",
+                'mandatory' => false,
+                'condition' => $this->validateDatabaseTimezoneSupport($conn),
+                'memo' => "{$name} should be configured with <a rel='noopener' target='_blank' href='{$tzUrl}'>full timezone support</a>.",
             );
         }
         break;
@@ -43,7 +60,7 @@ switch ($this->dbDriver) {
             'name' => 'PDO PostgreSQL extension',
             'mandatory' => true,
             'condition' => extension_loaded('pdo_pgsql'),
-            'memo' => 'The <https://secure.php.net/manual/en/ref.pdo-pgsql.php>PDO PostgreSQL</a> extension is required.'
+            'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/ref.pdo-pgsql.php">PDO PostgreSQL</a> extension is required.'
         );
         if ($conn !== false) {
             $requirements[] = array(
@@ -59,6 +76,7 @@ switch ($this->dbDriver) {
 // Only run this requirement check if we're running in the context of Craft.
 if (class_exists('Craft')) {
     $requirements[] = $this->webrootRequirement();
+    $requirements[] = $this->webAliasRequirement();
 }
 
 $requirements = array_merge($requirements, array(
@@ -66,92 +84,98 @@ $requirements = array_merge($requirements, array(
         'name' => 'Reflection extension',
         'mandatory' => true,
         'condition' => extension_loaded('reflection'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/class.reflectionextension.php">Reflection</a> extension is required.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/class.reflectionextension.php">Reflection</a> extension is required.',
     ),
     array(
         'name' => 'PCRE extension (with UTF-8 support)',
         'mandatory' => true,
         'condition' => extension_loaded('pcre') && preg_match('/./u', 'Ü') === 1,
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.pcre.php">PCRE</a> extension is required and it must be compiled to support UTF-8.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.pcre.php">PCRE</a> extension is required and it must be compiled to support UTF-8.',
     ),
     array(
         'name' => 'SPL extension',
         'mandatory' => true,
         'condition' => extension_loaded('SPL'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.spl.php">SPL</a> extension is required.'
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.spl.php">SPL</a> extension is required.'
+    ),
+    array(
+        'name' => 'BCMath extension',
+        'mandatory' => true,
+        'condition' => extension_loaded('bcmath'),
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.bc.php">BCMath</a> extension is required.'
     ),
     array(
         'name' => 'PDO extension',
         'mandatory' => true,
         'condition' => extension_loaded('pdo'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.pdo.php">PDO</a> extension is required.'
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.pdo.php">PDO</a> extension is required.'
     ),
     array(
         'name' => 'Multibyte String extension (with Function Overloading disabled)',
         'mandatory' => true,
         'condition' => extension_loaded('mbstring') && ini_get('mbstring.func_overload') == 0,
-        'memo' => 'Craft CMS requires the <a rel="noopener" target="_blank" href="http://www.php.net/manual/en/book.mbstring.php">Multibyte String</a> extension with <a rel="noopener" target="_blank" href="http://php.net/manual/en/mbstring.overload.php">Function Overloading</a> disabled in order to run.'
+        'memo' => 'Craft CMS requires the <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.mbstring.php">Multibyte String</a> extension with <a rel="noopener" target="_blank" href="https://php.net/manual/en/mbstring.overload.php">Function Overloading</a> disabled in order to run.'
     ),
     array(
         'name' => 'GD extension or ImageMagick extension',
-        'mandatory' => true,
+        'mandatory' => false,
         'condition' => extension_loaded('gd') || (extension_loaded('imagick') && !empty(\Imagick::queryFormats())),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.image.php">GD</a> or <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.imagick.php">ImageMagick</a> extension is required, however ImageMagick is recommended as it adds animated GIF support, and preserves 8-bit and 24-bit PNGs during image transforms.'
+        'memo' => 'When using Craft\'s default image transformer, the <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.image.php">GD</a> or <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.imagick.php">ImageMagick</a> extension is required. ImageMagick is recommended as it adds animated GIF support, and preserves 8-bit and 24-bit PNGs during image transforms.'
     ),
     array(
         'name' => 'GD extension',
         'mandatory' => false,
         'condition' => extension_loaded('gd'),
-        'memo' => '<a rel="noopener" target="_blank" href="http://php.net/manual/en/book.image.php">GD</a> extension is enabled.'
+        'memo' => '<a rel="noopener" target="_blank" href="https://php.net/manual/en/book.image.php">GD</a> extension is enabled.'
     ),
     array(
         'name' => 'ImageMagick extension',
         'mandatory' => false,
         'condition' => (extension_loaded('imagick') && !empty(\Imagick::queryFormats())),
-        'memo' => '<a rel="noopener" target="_blank" href="http://php.net/manual/en/book.imagick.php">ImageMagick</a> extension is enabled.'
+        'memo' => '<a rel="noopener" target="_blank" href="https://php.net/manual/en/book.imagick.php">ImageMagick</a> extension is enabled.'
     ),
     array(
         'name' => 'OpenSSL extension',
         'mandatory' => true,
         'condition' => extension_loaded('openssl'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.openssl.php">OpenSSL</a> extension is required.'
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.openssl.php">OpenSSL</a> extension is required.'
     ),
     array(
         'name' => 'cURL extension',
         'mandatory' => true,
         'condition' => extension_loaded('curl'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.curl.php">cURL</a> extension is required.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.curl.php">cURL</a> extension is required.',
     ),
     array(
         'name' => 'ctype extension',
         'mandatory' => true,
         'condition' => extension_loaded('ctype'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/book.ctype.php">ctype</a> extension is required.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.ctype.php">ctype</a> extension is required.',
     ),
     $this->iniSetRequirement(),
     array(
         'name' => 'Intl extension',
-        'mandatory' => false,
+        'mandatory' => true,
         'condition' => $this->checkPhpExtensionVersion('intl', '1.0.2', '>='),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://www.php.net/manual/en/book.intl.php">Intl</a> extension (version 1.0.2+) is recommended.'
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.intl.php">Intl</a> extension (version 1.0.2+) is recommended.'
     ),
     array(
         'name' => 'Fileinfo extension',
         'mandatory' => true,
         'condition' => extension_loaded('fileinfo'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.fileinfo.php">Fileinfo</a> extension required.'
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.fileinfo.php">Fileinfo</a> extension required.'
     ),
     array(
         'name' => 'DOM extension',
         'mandatory' => true,
         'condition' => extension_loaded('dom'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="http://php.net/manual/en/book.dom.php">DOM</a> extension is required.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.dom.php">DOM</a> extension is required.',
     ),
     array(
         'name' => 'iconv extension',
         'mandatory' => true,
         'condition' => function_exists('iconv'),
-        'memo' => '<a rel="noopener" target="_blank" href="http://php.net/manual/en/book.iconv.php">iconv</a> is required for more robust character set conversion support.',
+        'memo' => '<a rel="noopener" target="_blank" href="https://php.net/manual/en/book.iconv.php">iconv</a> is required for more robust character set conversion support.',
     ),
     $this->memoryLimitRequirement(),
     $this->maxExecutionTimeRequirement(),
@@ -159,61 +183,55 @@ $requirements = array_merge($requirements, array(
         'name' => 'password_hash()',
         'mandatory' => true,
         'condition' => function_exists('password_hash'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/function.password-hash.php">password_hash()</a> function is required so Craft can create secure passwords.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.password-hash.php">password_hash()</a> function is required so Craft can create secure passwords.',
     ),
     array(
         'name' => 'Zip extension',
         'mandatory' => true,
         'condition' => extension_loaded('zip'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/book.zip.php">zip</a> extension is required for zip and unzip operations.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.zip.php">zip</a> extension is required for zip and unzip operations.',
     ),
     array(
         'name' => 'JSON extension',
         'mandatory' => true,
         'condition' => extension_loaded('json'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/book.json.php">JSON</a> extension is required for JSON encoding and decoding.',
-    ),
-    array(
-        'name' => 'SOAP extension',
-        'mandatory' => true,
-        'condition' => extension_loaded('soap'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/book.soap.php">SOAP</a> extension is required to check VAT numbers.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/book.json.php">JSON</a> extension is required for JSON encoding and decoding.',
     ),
     array(
         'name' => 'proc_open()',
         'mandatory' => false,
         'condition' => function_exists('proc_open'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/function.proc-open.php">proc_open()</a> function is required for Plugin Store operations as well as sending emails.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.proc-open.php">proc_open()</a> function is required for Plugin Store operations as well as sending emails.',
     ),
     array(
         'name' => 'proc_get_status()',
         'mandatory' => false,
         'condition' => function_exists('proc_get_status'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/function.proc-get-status.php">proc_get_status()</a> function is required for Plugin Store operations as well as sending emails.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.proc-get-status.php">proc_get_status()</a> function is required for Plugin Store operations as well as sending emails.',
     ),
     array(
         'name' => 'proc_close()',
         'mandatory' => false,
         'condition' => function_exists('proc_close'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/function.proc-close.php">proc_close()</a> function is required for Plugin Store operations as well as sending emails.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.proc-close.php">proc_close()</a> function is required for Plugin Store operations as well as sending emails.',
     ),
     array(
         'name' => 'proc_terminate()',
         'mandatory' => false,
         'condition' => function_exists('proc_terminate'),
-        'memo' => 'The <a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/function.proc-terminate.php">proc_terminate()</a> function is required for Plugin Store operations as well as sending emails.',
+        'memo' => 'The <a rel="noopener" target="_blank" href="https://php.net/manual/en/function.proc-terminate.php">proc_terminate()</a> function is required for Plugin Store operations as well as sending emails.',
     ),
     array(
         'name' => 'allow_url_fopen',
         'mandatory' => false,
         'condition' => ini_get('allow_url_fopen'),
-        'memo' => '<a rel="noopener" target="_blank" href="https://secure.php.net/manual/en/filesystem.configuration.php#ini.allow-url-fopen">allow_url_fopen</a> must be enabled in your PHP configuration for Plugin Store and updating operations.',
+        'memo' => '<a rel="noopener" target="_blank" href="https://php.net/manual/en/filesystem.configuration.php#ini.allow-url-fopen">allow_url_fopen</a> must be enabled in your PHP configuration for Plugin Store and updating operations.',
     ),
     array(
         'name' => 'ignore_user_abort()',
         'mandatory' => false,
         'condition' => function_exists('ignore_user_abort'),
-        'memo' => '<a rel="noopener" target="_blank" href="https://www.php.net/manual/en/function.ignore-user-abort.php">ignore_user_abort()</a> must be enabled in your PHP configuration for the native web-based queue runner to work.',
+        'memo' => '<a rel="noopener" target="_blank" href="https://php.net/manual/en/function.ignore-user-abort.php">ignore_user_abort()</a> must be enabled in your PHP configuration for the native web-based queue runner to work.',
     ),
     array(
         'name' => 'Redis Extension',